move markdown type correction
